--- a/Applications/ApplicationsLib/ProjectData.cpp
+++ b/Applications/ApplicationsLib/ProjectData.cpp
@@ -39,7 +39,7 @@
 #include "ProcessLib/DeactivatedSubdomain.h"
 
 // PhreeqcIO
-#include "ChemistryLib/CreatePhreeqcIO.h"
+#include "ChemistryLib/CreateChemicalSolverInterface.h"
 #include "ProcessLib/ComponentTransport/ComponentTransportProcess.h"
 
 // FileIO
@@ -1032,6 +1032,9 @@ void ProjectData::parseChemicalSystem(
             ProcessLib::ComponentTransport::ComponentTransportProcess const*>(
             _processes[0].get()))
     {
+        auto const& process_id_to_component_name_map =
+            component_transport_process->getProcessIDToComponentNameMap();
+
         auto const chemical_solver =
             //! \ogs_file_attr{prj__chemical_system__chemical_solver}
             config->getConfigAttribute<std::string>("chemical_solver");
@@ -1040,20 +1043,30 @@ void ProjectData::parseChemicalSystem(
         {
             INFO(
                 "Configuring phreeqc interface for water chemistry "
-                "calculation.");
+                "calculation using file-based approach.");
 
-            auto const& process_id_to_component_name_map =
-                component_transport_process->getProcessIDToComponentNameMap();
+            _chemical_system = ChemistryLib::createChemicalSolverInterface<
+                ChemistryLib::ChemicalSolver::Phreeqc>(
+                *_mesh_vec[0], process_id_to_component_name_map, *config,
+                output_directory);
+        }
+        else if (boost::iequals(chemical_solver, "PhreeqcKernel"))
+        {
+            INFO(
+                "Configuring phreeqc interface for water chemistry "
+                "calculation by direct memory access.");
 
-            _chemical_system = ChemistryLib::createPhreeqcIO(
+            _chemical_system = ChemistryLib::createChemicalSolverInterface<
+                ChemistryLib::ChemicalSolver::PhreeqcKernel>(
                 *_mesh_vec[0], process_id_to_component_name_map, *config,
                 output_directory);
         }
         else
         {
             OGS_FATAL(
-                "Unknown chemical solver. Please specify Phreeqc as the solver "
-                "for water chemistry calculation instead.");
+                "Unknown chemical solver. Please specify either Phreeqc or "
+                "PhreeqcKernel as the solver for water chemistry calculation "
+                "instead.");
         }
     }
     else