--- a/ChemistryLib/ChemicalSolverType.h
+++ b/ChemistryLib/ChemicalSolverType.h
@@ -1,0 +1,19 @@
+/**
+ * \copyright
+ * Copyright (c) 2012-2019, OpenGeoSys Community (http://www.opengeosys.org)
+ *            Distributed under a Modified BSD License.
+ *              See accompanying file LICENSE.txt or
+ *              http://www.opengeosys.org/project/license
+ *
+ */
+
+#pragma once
+
+namespace ChemistryLib
+{
+enum class ChemicalSolver
+{
+    Phreeqc,
+    PhreeqcKernel
+};
+}  // namespace ChemistryLib