--- a/ChemistryLib/ChemicalSolverType.h
+++ b/ChemistryLib/ChemicalSolverType.h
@@ -1,4 +1,5 @@
 /**
+ * \file
  * \copyright
  * Copyright (c) 2012-2019, OpenGeoSys Community (http://www.opengeosys.org)
  *            Distributed under a Modified BSD License.