--- a/ChemistryLib/PhreeqcIOData/CreateAqueousSolution.cpp
+++ b/ChemistryLib/PhreeqcIOData/CreateAqueousSolution.cpp
@@ -8,6 +8,7 @@
  *
  */
 
+#include "CreateAqueousSolution.h"
 #include "AqueousSolution.h"
 #include "BaseLib/ConfigTree.h"
 #include "BaseLib/Error.h"