--- a/ChemistryLib/PhreeqcIOData/CreateAqueousSolution.h
+++ b/ChemistryLib/PhreeqcIOData/CreateAqueousSolution.h
@@ -10,6 +10,9 @@
 
 #pragma once
 
+#include <string>
+#include <vector>
+
 namespace BaseLib
 {
 class ConfigTree;
@@ -21,6 +24,9 @@ namespace PhreeqcIOData
 {
 struct AqueousSolution;
 
-AqueousSolution createAqueousSolution(BaseLib::ConfigTree const& config);
+AqueousSolution createAqueousSolution(
+    BaseLib::ConfigTree const& config,
+    std::vector<std::pair<int, std::string>> const&
+        process_id_to_component_name_map);
 }  // namespace PhreeqcIOData
 }  // namespace ChemistryLib