@joboog 2019-09-02
Is this replacing the existing OGS-6--PhreeqcInterface, or will there be an option to choose the interface type?