@renchao-lu 2019-09-02
The latter. Switching the tag `<chemical_solver>` from `Phreeqc` to `PhreeqcKernel` is meant to call the new phreeqc interface.