@montoyav 2019-09-03
Maybe interesting to know/mention how many cores were used for the mentioned simulations.