@renchao-lu 2019-09-03
> Maybe interesting to know/mention how many cores were used for the mentioned simulations.

Parallelization not touched in this PR. Just single core. @montoyav