@renchao-lu 2019-09-16
> There is some doxygen error resulting in 19 warnings -- should be fixed.
> As in #2660, please add `\file` in all headers -- there is a script in the PR's commits' comments.
> 
> There are some known deficiencies because of the way how Phreeqc is written, too long to describe here. I like that most of the Phreeqc calls are implemented in private functions and aren't mixed with the OGS code. Let's continue the journey to fast Phreeqc interface.

Thanks for the command.  Added `\file`.